+++ packages/data-provider/src/config/config.ts
@@ -1,10 +1,9 @@
 import { NodePgDatabase } from 'drizzle-orm/node-postgres';
-import { Request } from 'express';
 
 import type { DbConfig } from '@overture-stack/lyric-data-model';
 import * as schema from '@overture-stack/lyric-data-model/models';
 
-import type { UserSession } from '../utils/express.js';
+import type { AuthConfig } from '../middleware/auth.js';
 import { Logger } from './logger.js';
 
 export type AuditConfig = {
@@ -37,18 +36,6 @@ export type IdServiceConfig = {
 	useLocal: boolean;
 	customAlphabet: string;
 	customSize: number;
-};
-
-export type AuthStatus = 'authenticated' | 'no-auth' | 'invalid-auth';
-
-export type UserSessionResult = {
-	user?: UserSession;
-	authStatus: AuthStatus;
-};
-
-export type AuthConfig = {
-	enabled: boolean;
-	customAuthHandler?: (req: Request) => UserSessionResult;
 };
 
 /**
