--- packages/data-provider/src/utils/express.ts
@@ -1,3 +1,0 @@
-export type UserSession = {
-	username: string;
-};
